Refector PopLongPollingService
* Use enum PollingResult to replace int value
* Move PopLongPollingService to upper level
* Move wakeUp and polling method into PopLongPollingService
