+++ broker/src/main/java/org/apache/rocketmq/broker/longpolling/PopRequest.java
@@ -16,6 +16,7 @@
  */
 package org.apache.rocketmq.broker.longpolling;
 
+import io.netty.channel.ChannelHandlerContext;
 import java.util.Comparator;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicLong;
@@ -27,20 +28,24 @@ import io.netty.channel.Channel;
 public class PopRequest {
     private static final AtomicLong COUNTER = new AtomicLong(Long.MIN_VALUE);
 
-    private RemotingCommand remotingCommand;
-    private Channel channel;
-    private long expired;
-    private AtomicBoolean complete = new AtomicBoolean(false);
+    private final RemotingCommand remotingCommand;
+    private final ChannelHandlerContext ctx;
+    private final long expired;
+    private final AtomicBoolean complete = new AtomicBoolean(false);
     private final long op = COUNTER.getAndIncrement();
 
-    public PopRequest(RemotingCommand remotingCommand, Channel channel, long expired) {
-        this.channel = channel;
+    public PopRequest(RemotingCommand remotingCommand, ChannelHandlerContext ctx, long expired) {
+        this.ctx = ctx;
         this.remotingCommand = remotingCommand;
         this.expired = expired;
     }
 
     public Channel getChannel() {
-        return channel;
+        return ctx.channel();
+    }
+
+    public ChannelHandlerContext getCtx() {
+        return ctx;
     }
 
     public RemotingCommand getRemotingCommand() {
@@ -63,7 +68,7 @@ public class PopRequest {
     public String toString() {
         final StringBuilder sb = new StringBuilder("PopRequest{");
         sb.append("cmd=").append(remotingCommand);
-        sb.append(", channel=").append(channel);
+        sb.append(", ctx=").append(ctx);
         sb.append(", expired=").append(expired);
         sb.append(", complete=").append(complete);
         sb.append(", op=").append(op);
